EpicEvents/pyproject.toml
[tool.poetry]
name = "epicevents"
version = "0.1.0"
description = ""
authors = ["Kern84 <adrienleroy1984@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-filter = "^21.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
